requirements.txt
# Python 3.13
# Core
gymnasium>=1.0.0
numpy>=2.0.0
torch>=2.4.0

# RL
stable-baselines3>=2.4.0
sb3-contrib>=2.4.0

# Logging & visualization
tensorboard>=2.18.0
matplotlib>=3.9.0
tqdm>=4.67.0
rich>=13.0.0

# TensorBoard needs pkg_resources from setuptools<81
setuptools<81

# Rendering (optional — only needed for evaluate.py with --render)
pygame-ce>=2.5.0
